test(api): add scenario for successful recipe creation
Adds a Karate test for the happy path of the POST /recipes endpoint.

This test asserts that a successful creation with a valid token should return a 201 Created status code and include the new resource in the response. This test is expected to fail until the controller is updated in the next commit.